Added spacial methods to GameTile.

+++ src/model/interfaces/ResourceBehavior.java
@@ -1,5 +1,5 @@
 package src.model.interfaces;
 
 public interface ResourceBehavior {
-	public int harvest(float n);
+	public int harvest(int n);
 }
